+++ muxers/mplex/tests/async_write.rs
@@ -21,7 +21,7 @@
 use futures::{channel::oneshot, prelude::*};
 use libp2p::core::muxing::StreamMuxerExt;
 use libp2p::core::{upgrade, Transport};
-use libp2p::tcp::TcpTransport;
+use libp2p::tcp;
 
 #[test]
 fn async_write() {
@@ -32,7 +32,7 @@ fn async_write() {
     let bg_thread = async_std::task::spawn(async move {
         let mplex = libp2p::mplex::MplexConfig::new();
 
-        let mut transport = TcpTransport::default()
+        let mut transport = tcp::async_io::Transport::default()
             .and_then(move |c, e| upgrade::apply(c, mplex, e, upgrade::Version::V1))
             .boxed();
 
@@ -68,7 +68,7 @@ fn async_write() {
 
     async_std::task::block_on(async {
         let mplex = libp2p::mplex::MplexConfig::new();
-        let mut transport = TcpTransport::default()
+        let mut transport = tcp::async_io::Transport::default()
             .and_then(move |c, e| upgrade::apply(c, mplex, e, upgrade::Version::V1));
 
         let mut client = transport.dial(rx.await.unwrap()).unwrap().await.unwrap();