+++ transports/tcp/src/provider/tokio.rs
@@ -30,6 +30,8 @@ use std::net;
 use std::pin::Pin;
 use std::task::{Context, Poll};
 
+pub type Transport = crate::GenTcpTransport<Tcp>;
+
 #[derive(Copy, Clone)]
 pub enum Tcp {}